Player character with animation.
I managed to implement an animated player character that can walk around
(the potentially infinite) game world. Besides that, I also did some
cleaning up. The tile renderer is now less hacky, and I fixed the
hard-coded file paths, so that they are relative to the
asdf:source-file-directory.

Still not a game though.