Fixed tile-renderer.
Okay, as it turned out there were still a lot of bugs lurking. But they
have been squashed now. The tile renderer works correctly, and the
character is rendered in its proper layer so as to appear within the
scene. The character also stays in the center of the screen now, and the
"world revolves" around him, making it much more immersive and almost
giving you the impression that you are playing a game.

That last change turned out to be a source of a last-time error, whose
solution I realized while typing the message of a commit I intended to
do earlier. Namely, I simply rendered the tile map at `center-of-screen
- player-pos`, but `player-pos` is a `vec2` of floats. This lead to
rounding errors which resulted in white cracks showing between the
tiles.

So the tile-renderer is basically done now (even though I'm a bit
worried about the performance). It remains to

- complete the map (in particular, create the missing tiles)
- code ui
- code animations
- code gameplay

There are still two days and five hours left, so some hope too.